refactor(vllm): separate generic and engine refit APIs
Keep timing and future shared contracts under the engine-agnostic refit package while isolating vLLM placement and installation behind its engine adapter. Preserve existing imports through compatibility shims.
