+++ modelexpress_client/python/modelexpress/refit/__init__.py
@@ -1,13 +1,9 @@
 # SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Compatibility exports for :mod:`modelexpress.refit.timing`.
+"""Engine-agnostic primitives for live model refit."""
 
-New code should import timing primitives from ``modelexpress.refit`` or
-``modelexpress.refit.timing``.
-"""
-
-from .refit.timing import (
+from .timing import (
     MX_REFIT_TIMING_PREFIX,
     REFIT_TIMING_STAGES,
     RefitTimingRecorder,
