+++ modelexpress_client/python/tests/test_refit_timing.py
@@ -6,13 +6,16 @@ import logging
 
 import pytest
 
-from modelexpress.refit_timing import (
+from modelexpress.refit import (
     MX_REFIT_TIMING_PREFIX,
     REFIT_TIMING_STAGES,
     RefitTimingRecorder,
     current_refit_timing,
     refit_span,
     use_refit_timing,
+)
+from modelexpress.refit_timing import (
+    RefitTimingRecorder as CompatibilityRefitTimingRecorder,
 )
 
 
@@ -25,6 +28,10 @@ class _Clock:
 
     def advance(self, seconds):
         self.now += seconds
+
+
+def test_legacy_refit_timing_import_is_compatible():
+    assert CompatibilityRefitTimingRecorder is RefitTimingRecorder
 
 
 def test_stage_order_repeated_spans_and_unattributed_sum():
